# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "O2geeks Headless CMS",
    "description": "\n**Backend API for the O2geeks Branding System**\n\nThis API powers the decoupled content management backend. It provides robust RESTful endpoints for delivering and managing multi-format content.\n\n### Key Features:\n- **Authentication**: JWT-based secure access with Refresh Tokens.\n- **RBAC**: Strict Role-Based Access Control (Super Admin, Admin, Editor, User).\n- **Rate Limiting**: Distributed Redis-backed abuse protection.\n- **Media**: Dual-provider (Local + AWS S3) Multipart/Form-Data uploads.\n- **Engagement**: Zero-N+1 likes, comments, and favorites.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/webhooks": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Create Webhook",
        "description": "Register a new webhook.",
        "operationId": "create_webhook_api_v1_webhooks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreateOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "List Webhooks",
        "description": "List all registered webhooks.",
        "operationId": "list_webhooks_api_v1_webhooks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WebhookOut_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/{webhook_id}": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Get Webhook",
        "description": "Get a specific webhook by ID.",
        "operationId": "get_webhook_api_v1_webhooks__webhook_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Webhook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Update Webhook",
        "description": "Update a webhook.",
        "operationId": "update_webhook_api_v1_webhooks__webhook_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Webhook Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Delete Webhook",
        "description": "Delete a webhook.",
        "operationId": "delete_webhook_api_v1_webhooks__webhook_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Webhook Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/{webhook_id}/logs": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Get Webhook Logs",
        "description": "View delivery logs for a specific webhook.",
        "operationId": "get_webhook_logs_api_v1_webhooks__webhook_id__logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Webhook Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WebhookLogOut_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/webhooks/{webhook_id}/test": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Test Webhook",
        "description": "Send a test ping to the webhook endpoint.",
        "operationId": "test_webhook_api_v1_webhooks__webhook_id__test_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Webhook Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookTestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_WebhookLogOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WebhookLogOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[WebhookLogOut]"
      },
      "PaginatedResponse_WebhookOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WebhookOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[WebhookOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookCreate": {
        "properties": {
          "url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Url",
            "description": "Target endpoint URL (must be a valid HTTP/HTTPS URL)"
          },
          "event": {
            "type": "string",
            "title": "Event",
            "description": "Event name, e.g. content.published"
          },
          "content_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Types",
            "description": "List of content types to trigger on, e.g. [\"blog\", \"news\"] or [\"*\"] for all"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Human-readable label for this webhook"
          }
        },
        "type": "object",
        "required": [
          "url",
          "event",
          "content_types"
        ],
        "title": "WebhookCreate"
      },
      "WebhookCreateOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "content_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Types"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "event",
          "content_types",
          "description",
          "is_active",
          "created_at",
          "updated_at",
          "secret"
        ],
        "title": "WebhookCreateOut"
      },
      "WebhookLogOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "webhook_id": {
            "type": "integer",
            "title": "Webhook Id"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "content_id": {
            "type": "integer",
            "title": "Content Id"
          },
          "request_url": {
            "type": "string",
            "title": "Request Url"
          },
          "response_status": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Status"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "request_body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Body"
          },
          "response_body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Body"
          },
          "delivered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Delivered At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "webhook_id",
          "event",
          "content_type",
          "content_id",
          "request_url",
          "response_status",
          "success",
          "error_message",
          "request_body",
          "response_body",
          "delivered_at"
        ],
        "title": "WebhookLogOut"
      },
      "WebhookOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "event": {
            "type": "string",
            "title": "Event"
          },
          "content_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Content Types"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "secret": {
            "type": "string",
            "title": "Secret"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "event",
          "content_types",
          "description",
          "is_active",
          "created_at",
          "updated_at",
          "secret"
        ],
        "title": "WebhookOut"
      },
      "WebhookTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "status_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Code"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "WebhookTestResponse"
      },
      "WebhookUpdate": {
        "properties": {
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "event": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event"
          },
          "content_types": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Types"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "WebhookUpdate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}